con reglas en frontend

+++ src/stores/database.js
@@ -1,35 +1,35 @@
 import { defineStore } from 'pinia'
 import { query, 
-        collection, 
-        getDocs,
-        doc,
-        addDoc,
-        deleteDoc,
-        getDoc,
-        where,
-        updateDoc} from 'firebase/firestore/lite'
-import {db, auth} from '../firebaseConfig'
+         collection, 
+         getDocs,
+         doc,
+         addDoc,
+         deleteDoc,
+         getDoc,
+         where,
+         updateDoc } from 'firebase/firestore/lite'
+import { db, auth } from '../firebaseConfig'
 import { nanoid } from 'nanoid'
 import router from '../router'
 
-
-
 export const useDatabaseStore = defineStore( 'urls', {
+
   state: ()=>({
     documents: [], //un array de objetos de momento vacio
     loadingDocs:  false
   }),
+
   actions: {
     async getUrls(){
       try{
-
         if(this.documents.length != 0){
           return
         }
         this.loadingDocs = true
         // filtramos los docs de un usuario con where
         const q = query(collection(db, 'urls'), 
-          where( "user", "==", auth.currentUser.uid))
+           where( "user", "==", auth.currentUser.uid)
+          )
         const querySnapShot = await getDocs(q)        
         querySnapShot.forEach(doc => this.documents.push(
           // llenamos documents con este objeto
@@ -49,13 +49,11 @@ export const useDatabaseStore = defineStore( 'urls', {
 
     async addUrl(name){
       try{
-
         const newDoc = {
           name: name,
           short: nanoid(6),
           user: auth.currentUser.uid
         }
-
         const docRef = await addDoc(collection(db, "urls"), newDoc);
         console.log("Document written with ID: ", docRef.id); 
          // añadimos a documents este objeto
@@ -74,19 +72,37 @@ export const useDatabaseStore = defineStore( 'urls', {
     async leerUrl(id){
       try{
         const docRef = doc( db, 'urls', id)       
-        const docSnap = await getDoc( docRef)  
+        const docSnap = await getDoc( docRef) 
+        
+        // reglas de seguridad en el frontend
+        if( !docSnap.exists){
+          throw new Error ('El documento no existe!')
+        }
+        if( docSnap.data().user != auth.currentUser.uid){
+          throw new Error ('El documento no le pertenece!')
+        }
+
         // console.log( docSnap.data().name)
         return docSnap.data().name
       }
       catch(err){
-        console.log(err)
+        console.log(err.msg)
       }
-
     },
 
     async deleteUrl(id){
       try{
         const docRef = doc( db, 'urls', id)
+        const docSnap = await getDoc( docRef) 
+
+         // reglas de seguridad en el frontend
+        if( !docSnap.exists){
+          throw new Error ('El documento no existe!')
+        }
+        if( docSnap.data().user != auth.currentUser.uid){
+          throw new Error ('El documento no le pertenece!')
+        }
+
         await deleteDoc(docRef);
         console.log('Eliminado de la db') 
         // lo eliminamos del store
@@ -98,22 +114,31 @@ export const useDatabaseStore = defineStore( 'urls', {
       finally{
       }
     },
+
     async updateUrl(id, name){
       try{
         const docRef = doc( db, 'urls', id)
+        const docSnap = await getDoc( docRef) 
+
+         // reglas de seguridad en el frontend
+        if( !docSnap.exists){
+          throw new Error ('El documento no existe!')
+        }
+        if( docSnap.data().user != auth.currentUser.uid){
+          throw new Error ('El documento no le pertenece!')
+        }
+        
         await updateDoc( docRef, {name: name}) 
         console.log('Actualizado en la db') 
         // actualizamos tambien el store
         this.documents = this.documents.map( item => item.id == id ? ({...item, name: name}) : item) 
-        router.push('/')
-        // router.push('/')
+        router.push('/')       
       }
       catch(err){
         console.log(err)
       }
       finally{        
       }
-    }
-    
+    }    
   }
 })
